* Fixed bug where iterating over a folder that may not exist

+++ Modules/Cache/Libraries/Cache.php
@@ -95,9 +95,12 @@ class Cache
 	{
 		$filenames = glob( $folder . $this->cacheID . "--*" );
 
-		foreach( $filenames as $filename )
+		if( $filenames !== false )
 		{
-			unlink( $filename );
+			foreach( $filenames as $filename )
+			{
+				unlink( $filename );
+			}
 		}
 	}
